+++ src/app.ts
@@ -4,6 +4,7 @@ import { CommandsSubscriber } from './subscribers/commands.js';
 import { ConflictsNotifierSubscriber } from './subscribers/conflicts-notifier.js';
 import { IssueIntakeSubscriber } from './subscribers/issue-intake.js';
 import { LockOldIssuesSubscriber } from './subscribers/lock-old-issues.js';
+import { MilestoneSubscriber } from './subscribers/milestone.js';
 import { NoMergeCommitsSubscriber } from './subscribers/no-merge-commits.js';
 import { NoResponseCloserSubscriber } from './subscribers/no-response-closer.js';
 import { PrTitleLinterSubscriber } from './subscribers/pr-title-linter.js';
@@ -22,6 +23,7 @@ export const carson = new Carson([
   new ConflictsNotifierSubscriber(),
   new IssueIntakeSubscriber(),
   new LockOldIssuesSubscriber(),
+  new MilestoneSubscriber(),
   new NoMergeCommitsSubscriber(),
   new NoResponseCloserSubscriber(),
   new PrTitleLinterSubscriber(),